refactor(LineMaker.java): LineMaker 인터페이스 삭제
