+++ src/main/java/nextstep/ladder/App.java
@@ -2,7 +2,7 @@ package nextstep.ladder;
 
 import nextstep.ladder.domain.Ladder;
 import nextstep.ladder.domain.LadderMaker;
-import nextstep.ladder.domain.LineMakerImpl;
+import nextstep.ladder.domain.LineMaker;
 import nextstep.ladder.domain.Users;
 import nextstep.ladder.ui.InputView;
 import nextstep.ladder.ui.ResultView;
@@ -10,7 +10,7 @@ import nextstep.ladder.ui.ResultView;
 public class App {
     public static void main(String[] args) {
         Users users = new Users(InputView.getUserNames());
-        Ladder ladder = new LadderMaker(new LineMakerImpl(users.size())).makeLadder(InputView.getLadderHeight());
+        Ladder ladder = new LadderMaker(new LineMaker(users.size())).makeLadder(InputView.getLadderHeight());
         ResultView.printResult(users, ladder);
     }
 }
