+++ src/test/java/nextstep/ladder/domain/LadderMakerTest.java
@@ -9,7 +9,7 @@ public class LadderMakerTest {
     @Test
     @DisplayName("사다리 높이가 6이면 라인수는 6개이다")
     void size() {
-        LadderMaker ladderMaker = new LadderMaker(new LineMakerImpl(5));
+        LadderMaker ladderMaker = new LadderMaker(new LineMaker(5));
         assertThat(ladderMaker.makeLadder(6).getLines()).hasSize(6);
     }
 }
